Do not log initialization info in prod
This change prevents loggin the initialization info
in prod environment, as the logging state is handled
in memory and serverless environments will log it on
every invocation of the docs route handlers.

This also refactors the project structure so that app
router and pages router specific code is split into
separate folders.
