+++ packages/next-rest-framework/src/app-router/index.ts
@@ -1,0 +1,3 @@
+export * from './docs-route-handler';
+export * from './route-handler';
+export * from './route-operation';
