+++ packages/next-rest-framework/src/index.ts
@@ -1,12 +1,11 @@
 import { NextResponse } from 'next/server';
 import { type TypedNextResponse as TypedNextResponseType } from './types';
 export {
-  routeHandler,
-  routeOperation,
+  docsApiRouteHandler,
   apiRouteHandler,
   apiRouteOperation
-} from './route-handlers';
-export { docsRouteHandler, docsApiRouteHandler } from './docs-handlers';
+} from './pages-router';
+export { docsRouteHandler, routeHandler, routeOperation } from './app-router';
 
 // @ts-expect-error - Keep the original NextResponse functionality with custom types.
 export const TypedNextResponse: typeof TypedNextResponseType = NextResponse;
