+++ packages/next-rest-framework/src/pages-router/index.ts
@@ -1,0 +1,3 @@
+export * from './api-route-handler';
+export * from './api-route-operation';
+export * from './docs-api-route-handler';
