+++ packages/next-rest-framework/src/utils/index.ts
@@ -1,3 +1,5 @@
 export * from './config';
+export * from './docs';
+export * from './logging';
 export * from './open-api';
 export * from './schemas';
